+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### `jsonschema-module-jackson`
+#### Fixed
+- support `@JacksonAnnotationsInside` annotated combo annotations also when looking for `@JsonPropertyDescription`  
 
 ## [4.38.0] - 2025-03-24
 ### `jsonschema-generator`